pyproject.toml
[tool.poetry]
name = "alugaveiculos-project"
version = "0.1.0"
description = "Sistema backend para alugação de veiculos."
authors = ["Rip4568 <63131796+Rip4568@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "alugaveiculos_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django-cors-headers = "^3.13.0"
django = "^4.1.5"
whitenoise = "^6.3.0"
dj-database-url = "^1.2.0"
python-dotenv = "^0.21.0"
djangorestframework = "^3.14.0"
dj-rest-auth = "^2.2.6"
django-allauth = "^0.52.0"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
